flask/requirements.txt
Flask==2.1.1
Flask-Cors==3.0.10
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.4.4
mysqlclient==2.0.3
SQLAlchemy==1.3.23
waitress==2.0.0
Werkzeug==2.0
PyMySQL==1.0.2
pandas==1.3.1
openpyxl==3.0.7
cryptography== 3.4.7
Flask-APScheduler==1.12.2
ncclient==0.6.10
xmltodict==0.12.0
influxdb-client==1.20.0
netmiko==3.4.0
textfsm==1.1.0
requests==2.22.0
flasgger==0.9.5
pysnmp==4.4.12
pyOpenSSL==21.0.0
ttp==0.8.2
PyJWT==2.3.0
itsdangerous==2.0.1
markupsafe==2.0.1
pysftp==0.2.9
netaddr==0.8.0
python-ldap==3.4.0
mac-vendor-lookup==0.1.12
